🐛 fix(ItemService.cs): add missing await to AddScriptableGameObjectAction method call ✨ feat(ItemService.cs): add support for scriptable game object actions ✨ feat(ItemsScriptModule.cs): add script functions to add game object actions and scheduled game object actions 🐛 fix(RaceObjectSeedEntity.cs): change TileType property type from int to string 🐛 fix(RaceEntity.cs): change TileId property type from int to uint 🐛 fix(PlayerDataHelper.cs): cast TileId to int in AddRace method 🐛 fix(DarkSunEngine.cs): uncomment network client disconnect call
🐛 fix(SeedService.cs): cast TileId to uint
✨ feat(SeedService.cs): add race seed data
✨ feat(TypeService.cs): add fallback for tile search
🐛 fix(WorldService.cs): return false if player is null in MovePlayerAsync

+++ DarkStar.Engine/MessageListeners/Helpers/PlayerDataHelper.cs
@@ -49,7 +49,7 @@ public static class PlayerDataHelper
                         Strength = r.Strength,
                         Name = r.Name,
                         Luck = r.Luck,
-                        TileId = r.TileId
+                        TileId = (int)r.TileId
                     }
                 )
                 .ToList()
